chore: rename mapping to question
