+++ collab-service/src/redis_adapter/redis_question_adapter.ts
